// types/schema.ts
export type IPlan = {
  id?: string;
  idIndex?: number;
  userId?: string;

  clientName?: string;
  videoThemeColor?: string;
  videoAvatar?: string;
  videoBackgroundMusic?: string;
  videoBackgroundImage?: File | string | null;
  buildSpanishVideo?: boolean;
  clientLogo?: File | string | null;

  // detail
  match?: string;
  planType?: string;
  matchPlan?: string;
  matchSafe?: string;
  nonElective?: string;
  entryDates?: string;
  showAdvancedInvestment?: boolean;
  showAdvancedDeferrals?: boolean;
  showAdvancedChildDeferrals?: boolean;
  advancedEntryDates?: string[];
  advancedDeferrals?: string[];
  investments?: string[];
  mandatoryContribution?: string;
  deferrals?: string[];
  advancedInvestments?: string[];
  ageRequirement?: string;
  vestingScheduleRadio?: string;
  vestingSchedules?: string[];
  employerContribution?: string;
  automaticEnrollment?: boolean;
  automaticIncrease?: boolean;
  customEntryDates?: string;
  customEntryDateType?: string;
  customEntryDatesValue?: string;
  fullCustomEntryDates?: string;
  advancedEntryHours?: string;
  matchType?: string;
  matchPercentage?: number;
  safeHarborMatch?: string;
  safeHarborMatchType?: string;
  safeHarborContribution?: string | number;
  showWaitingPeriod?: boolean;
  nonElectiveEmployerContributions?: boolean;
  employerProfitSharingContributions?: boolean;
  waitingPeriod?: boolean;
  automaticEnrollmentPercentage?: string;
  automaticEnrollmentWaitPeriod?: string;
  automaticIncreasePercentage?: string;
  automaticIncreaseCap?: string;
  advancedEntryDatesValue?: string;
  waitingPeriodDuration?: string;
  waitingPeriodStart?: string;
  waitingPeriodStartDate?: string;
  nonElectiveType?: string;
  nonElectivePercentage?: string;
  profitSharingType?: string;
  profitSharingPercentage?: string;
  useCustomText?: boolean;
  customText?: string;
  useProfitSharingCustomText?: boolean;
  profitSharingCustomText?: string;
  customScheduleYears?: string;
  customStructureText?: string;

  // resource
  script?: string;
  recordKeeperName?: string;
  recordKeeperPhone?: string;
  recordKeeperWebsite?: string;
  recordkeeper?: string;
  onlineEnrollment?: string;
  isDisplayRecodeKeeper?: boolean;
  title?: string;
  companyName?: string;
  contactName?: string;
  email?: string;
  phoneNumber?: string;
  planAdvisor?: string;
  companyContact?: string;
  tpa?: string;
  educationalVideos?: string;
  providerName?: string;
  providerLogo?: string;
  website?: string;
  providerPhoneNumber?: string;
  displayAdvisorInfoHeader?: boolean;
  tpaName?: string;
  tpaEmail?: string;
  tpaPhoneNumber?: string;
  planDocumentsLinks?: string;
  recordKeeperId?: string;
  addressCode?: string;
};